Decouple ArticleViewer shell from WhoColor authorship feature
## Changes

Splits the 462-line `ArticleViewer.jsx` (which entangled a generic article
viewer with WhoColor authorship highlighting) into a reusable, annotation-
agnostic shell plus a self-contained authorship feature, so a future consumer
(e.g. claim verification) can reuse the shell with its own highlight hook
without importing, touching, or seeing any authorship code.

The shell (zero authorship coupling):
- `containers/ArticleViewer.jsx` — now a 14-line public wrapper that injects the
  authorship hook; all four existing import sites are untouched.
- `containers/ArticleViewerShell.jsx` (new) — visibility, `?showArticle=`
  permalink, click-outside, parsed-HTML fetch + redirect cross-check, revision
  toggle, bad-work alert, footer chrome. Calls an injected `useHighlightFeature`
  hook (default no-op → plain HTML, no legend).
- `components/ParsedArticle.jsx` (single `html` prop), `components/Footer.jsx`
  (generic `legend` node slot), `utils/URLBuilder.js` + `utils/ArticleViewerAPI.js`
  (parsed-article only).

The authorship feature (self-contained):
- `hooks/useAuthorshipHighlighting.js` (new) — all WhoColor state + orchestration
  (`highlightAuthors` ported verbatim, user-id resolution, unhighlighted-
  contributor reconciliation, legend, status); returns
  `{ html, legend, buttonLabel, pending }`.
- `authorship/` — `ArticleViewerLegend.jsx`, `ArticleScroll.js`, `colors.js`
  (moved via git mv), plus new `AuthorshipURLBuilder.js` / `AuthorshipAPI.js`
  that extend the shell's builder/API. `article_graphs.jsx` colors import updated.

The seam uses imperative-timed signals, not effect-reaction guesses, to preserve
ordering: the shell bumps `parsedSettle {id, ok, message}` on each parsed-fetch
settle so WhoColor fires once, post-redirect-resolution; exposes a two-way
`requestTitleVerification`; and passes `article` by reference so the in-place
redirect-title mutation reaches the WhoColor fetch.

Intentional, user-invisible fixes confirmed in-scope by the user:
- Collapse the redundant double WhoColor fetch on revision-toggle/recovery.
- Revision button no longer stuck disabled (`pending` now clears on settle;
  previously stuck forever for non-WhoColor langs and after any WhoColor failure).
- Consolidate three overlapping user-id triggers into two guarded effects,
  closing a latent gap where IDs weren't fetched when `users` arrived after open.
- Remove confirmed-dead `__whocolorStatus` so the base API is authorship-free.

Specs: `URLBuilder.spec.js` keeps the parsed/wiki cases; whocolor + user-query
cases move to new `authorship/AuthorshipURLBuilder.spec.js`; new
`authorship/AuthorshipAPI.spec.js`. In `spec/features/article_viewer_spec.rb`,
fix a stale selector (`find('a', ...)` → `find('button', ...)`) for the
"Quality Problems?" button, which renders a `<button>`; this surfaced only when
the (otherwise pending) example was run live.

## Process

The refactor was implemented entirely in this session, guided by a spec document
authored by a previous session; no implementation code predated this conversation.

The user's direction was terse — roughly six messages, mostly a sentence or two:
the initial pointer to the spec, a green-light that fixing genuinely
redundant/buggy control flow is in-scope (given confirmation it's safe), a
request to enable the feature spec for local validation, and an instruction to
revert that enablement afterward while keeping the selector fix. The remaining
design judgment was delegated.

It was an extended, multi-phase exchange. I planned in plan mode and used a Plan
agent to pressure-test the shell↔feature seam before writing code — that critique
reshaped the design (imperative callbacks over state-signals; hook owns the
pending lifecycle; preserve `highlightAuthors` byte-for-byte). After implementing,
I enabled the feature spec for a one-off live run against the real MediaWiki and
WhoColor APIs: both endpoints send `access-control-allow-origin: *`, confirming
the long-standing "CORS error" pending note is CI-network-specific. Example 1
passed live; example 2 failed once on the stale `<a>` selector (diagnosed as a
pre-existing test bug, not a regression, since the example never actually ran),
which I fixed before reverting the temporary opt-in gating.

Tests: JS suite was green throughout — targeted Jest on the ArticleViewer specs
(24 examples), then the full `yarn test` twice (398 examples), with ESLint
(--max-warnings=0) and `yarn build` clean. The feature spec ran live three times
(example 1 pass; example 2 fail→fix→pass) plus a default run confirming the
unchanged "2 pending, 0 failures"; RuboCop clean.

(Commit message written by Claude Code.)
