+++ app/assets/javascripts/components/common/ArticleViewer/authorship/ArticleScroll.js
