+++ app/assets/javascripts/components/common/ArticleViewer/authorship/colors.js
