+++ app/assets/javascripts/components/common/ArticleViewer/components/ParsedArticle.jsx
@@ -5,12 +5,10 @@ import PropTypes from 'prop-types';
 const httpLinkMatcher = /(<a href="http)/g;
 const blankTargetLink = '<a target="_blank" href="http';
 
-export const ParsedArticle = ({ highlightedHtml, whocolorHtml, parsedArticle }) => {
-  let articleHTML = highlightedHtml || whocolorHtml || parsedArticle;
-
+export const ParsedArticle = ({ html }) => {
   // This sets `target="_blank"` for all of the non-anchor links in the article HTML,
   // so that clicking one will open it in a new tab.
-  articleHTML = articleHTML?.replace(httpLinkMatcher, blankTargetLink);
+  const articleHTML = html?.replace(httpLinkMatcher, blankTargetLink);
 
   return (
     <div className="parsed-article" dangerouslySetInnerHTML={{ __html: articleHTML }} />
@@ -18,9 +16,7 @@ export const ParsedArticle = ({ highlightedHtml, whocolorHtml, parsedArticle }) 
 };
 
 ParsedArticle.propTypes = {
-  highlightedHtml: PropTypes.string,
-  whocolorHtml: PropTypes.string,
-  parsedArticle: PropTypes.string
+  html: PropTypes.string
 };
 
 export default ParsedArticle;
